Fix issue #9.
Variable in settings treated as string in code but actually was an integer. After/before launcher functions better documented. xbmc.audioResume() is only called if Kodi was playing media: this speed ups a lot when returning from the launched program. If xbmc.audioResume() is called and audio was not previously suspended Kodi pauses for 2/3 seconds (on Linux).